+++ src/GameSrc/objsim.c
@@ -1268,7 +1268,7 @@ errtype obj_create_player(ObjLoc *plr_loc) {
     pos_list[0] = fix_from_obj_coord(plr_loc->x) >> 8;
     pos_list[1] = fix_from_obj_coord(plr_loc->y) >> 8;
     pos_list[2] = fix_from_obj_height_val(plr_loc->z) >> 8;
-    fr_objslew_go_real_height(NULL, (long *)pos_list);
+    fr_objslew_go_real_height(NULL, (int32_t *)pos_list);
     plr_loc->z = obj_height_from_fix(pos_list[2] << 8);
 
     if ((player_struct.edms_state[0]) && (!global_fullmap->cyber)) {
